Finished upgrad to sdk 23
